requirements.txt
streamlit
requests
pydub
python-dotenv
